Separate unit test constants by domain models.
