+++ src/test/kotlin/hu/netsurf/erp/CommonTestConstants.kt
@@ -1,0 +1,5 @@
+package hu.netsurf.erp
+
+object CommonTestConstants {
+    const val EMPTY_STRING = ""
+}
