+++ src/test/kotlin/hu/netsurf/erp/PhotoTestConstants.kt
@@ -1,0 +1,11 @@
+package hu.netsurf.erp
+
+object PhotoTestConstants {
+    const val MULTIPART_FILE_SIZE = 10485760
+    const val CONTENT_TYPE_IMAGE_JPEG = "image/jpeg"
+    const val ORIGINAL_FILE_NAME = "file_name.jpeg"
+    const val INVALID_ORIGINAL_FILE_NAME = "file_name.txt"
+    const val PHOTO_FILE_NAME = "7a759fbb-39d8-4b3b-af57-4266980901dc.jpeg"
+    const val UPLOADS_DIRECTORY_WITH_PHOTOS_SUBDIRECTORY_AND_CUSTOM_SUBDIRECTORY = "uploads/photos/products/"
+    val ALLOWED_EXTENSIONS = listOf("jpg", "jpeg", "png", "bmp")
+}
